refactor : s3 구조 변경, yml 추가, 오류 수정
