+++ meerket/meerket-infrastructure/src/main/resources/application-infrastructure.yml
@@ -11,3 +11,12 @@ spring:
         client_secret: ${NAVER_OAUTH_CLIENT_SECRET}
         state: ${NAVER_OAUTH_STATE}
 
+cloud:
+  aws:
+    credentials:
+      access-key: ${AWS_ACCESS_KEY_ID}
+      secret-key: ${AWS_ACCESS_KEY}
+    region:
+      static: ap-northeast-2
+    s3:
+      bucket-name: ${AWS_S3_BUCKET_NAME}
